+++ src/main/sono/io/Input.java
@@ -1,8 +1,6 @@
 package main.sono.io;
 
-import java.math.BigDecimal;
-
 public abstract class Input {
 	public abstract String getLine();
-	public abstract BigDecimal getNumber();
+	public abstract double getNumber();
 }
